Implement Actions Available -add action available to the character to the combat section for normal and SBL characters -fix error in starting wealth display

+++ app/src/main/java/com/paetus/animaCharCreator/activities/fragments/dialogs/DetailAlert.kt
@@ -1465,9 +1465,9 @@ fun EquipmentDetails(
     //create price string of the object
     val priceString =
             when(equipment.coinType){
-                CoinType.Gold -> stringResource(id = R.string.goldLabel, equipment.baseCost)
-                CoinType.Silver -> stringResource(id = R.string.silverLabel, equipment.baseCost)
-                CoinType.Copper -> stringResource(id = R.string.copperLabel, equipment.baseCost)
+                CoinType.Gold -> stringResource(id = R.string.goldLabel, equipment.baseCost.toInt())
+                CoinType.Silver -> stringResource(id = R.string.silverLabel, equipment.baseCost.toInt())
+                CoinType.Copper -> stringResource(id = R.string.copperLabel, equipment.baseCost.toInt())
             }
 
     Column{
